requirements.txt
django==3.0.2
django-filter==2.2.0
djangorestframework==3.11.0
django-rest-auth==0.9.5
